# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/menu-items": {
      "get": {
        "operationId": "MenuItemController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지당 항목 수",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "storeId",
            "required": false,
            "in": "query",
            "description": "매장 ID로 필터링",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "카테고리 ID로 필터링",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isAvailable",
            "required": false,
            "in": "query",
            "description": "판매 가능 여부로 필터링",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isPopular",
            "required": false,
            "in": "query",
            "description": "인기 메뉴 여부로 필터링",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isNew",
            "required": false,
            "in": "query",
            "description": "신메뉴 여부로 필터링",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isRecommended",
            "required": false,
            "in": "query",
            "description": "추천 메뉴 여부로 필터링",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "메뉴 아이템 목록을 반환합니다."
          }
        },
        "summary": "메뉴 아이템 목록 조회",
        "tags": [
          "메뉴 아이템"
        ]
      }
    },
    "/api/menu-items/{id}": {
      "get": {
        "operationId": "MenuItemController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "메뉴 아이템 상세 정보를 반환합니다."
          },
          "404": {
            "description": "메뉴 아이템을 찾을 수 없습니다."
          }
        },
        "summary": "메뉴 아이템 상세 조회",
        "tags": [
          "메뉴 아이템"
        ]
      },
      "patch": {
        "operationId": "MenuItemController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "메뉴 아이템이 수정되었습니다."
          },
          "403": {
            "description": "해당 메뉴에 대한 권한이 없습니다."
          },
          "404": {
            "description": "메뉴 아이템을 찾을 수 없습니다."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "메뉴 아이템 수정",
        "tags": [
          "메뉴 아이템"
        ]
      },
      "delete": {
        "operationId": "MenuItemController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "메뉴 아이템이 삭제되었습니다."
          },
          "403": {
            "description": "해당 메뉴에 대한 권한이 없습니다."
          },
          "404": {
            "description": "메뉴 아이템을 찾을 수 없습니다."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "메뉴 아이템 삭제",
        "tags": [
          "메뉴 아이템"
        ]
      }
    },
    "/api/menu-items/store/{storeId}": {
      "get": {
        "operationId": "MenuItemController_findByStoreId",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "매장별 메뉴 아이템 목록을 반환합니다."
          }
        },
        "summary": "매장별 메뉴 아이템 조회",
        "tags": [
          "메뉴 아이템"
        ]
      },
      "post": {
        "operationId": "MenuItemController_create",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateMenuItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "메뉴 아이템이 생성되었습니다."
          },
          "403": {
            "description": "해당 매장에 대한 권한이 없습니다."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "메뉴 아이템 생성",
        "tags": [
          "메뉴 아이템"
        ]
      }
    },
    "/api/menu-items/category/{categoryId}": {
      "get": {
        "operationId": "MenuItemController_findByCategoryId",
        "parameters": [
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "카테고리별 메뉴 아이템 목록을 반환합니다."
          }
        },
        "summary": "카테고리별 메뉴 아이템 조회",
        "tags": [
          "메뉴 아이템"
        ]
      }
    },
    "/api/menu-items/popular/store/{storeId}": {
      "get": {
        "operationId": "MenuItemController_findPopularItems",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "매장의 인기 메뉴를 반환합니다."
          }
        },
        "summary": "매장 인기 메뉴 조회",
        "tags": [
          "메뉴 아이템"
        ]
      }
    },
    "/api/menu-items/new/store/{storeId}": {
      "get": {
        "operationId": "MenuItemController_findNewItems",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "매장의 신메뉴를 반환합니다."
          }
        },
        "summary": "매장 신메뉴 조회",
        "tags": [
          "메뉴 아이템"
        ]
      }
    },
    "/api/menu-items/recommended/store/{storeId}": {
      "get": {
        "operationId": "MenuItemController_findRecommendedItems",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "매장의 추천 메뉴를 반환합니다."
          }
        },
        "summary": "매장 추천 메뉴 조회",
        "tags": [
          "메뉴 아이템"
        ]
      }
    },
    "/api/menu-items/{id}/stock": {
      "patch": {
        "operationId": "MenuItemController_updateStock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "메뉴 아이템 재고가 수정되었습니다."
          },
          "403": {
            "description": "해당 메뉴에 대한 권한이 없습니다."
          },
          "404": {
            "description": "메뉴 아이템을 찾을 수 없습니다."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "메뉴 아이템 재고 수정",
        "tags": [
          "메뉴 아이템"
        ]
      }
    },
...
    "/api/options/items/{id}": {
      "post": {
        "operationId": "OptionController_updateOptionItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "옵션 항목 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOptionItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "옵션 항목 수정 성공"
          },
          "404": {
            "description": "옵션 항목을 찾을 수 없음"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "옵션 항목 수정",
        "tags": [
          "옵션"
        ]
      },
      "delete": {
        "operationId": "OptionController_deleteOptionItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "옵션 항목 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "옵션 항목 삭제 성공"
          },
          "404": {
            "description": "옵션 항목을 찾을 수 없음"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "옵션 항목 삭제",
        "tags": [
          "옵션"
        ]
      }
    },
    "/api/options/menu-option-groups/menu/{menuId}": {
      "get": {
        "operationId": "OptionController_findMenuOptionGroupsByMenuId",
        "parameters": [
          {
            "name": "menuId",
            "required": true,
            "in": "path",
            "description": "메뉴 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "메뉴 옵션 그룹 목록 반환 성공"
          },
          "404": {
            "description": "메뉴를 찾을 수 없음"
          }
        },
        "summary": "메뉴에 할당된 옵션 그룹 조회",
        "tags": [
          "옵션"
        ]
      }
    },
...
  },
  "info": {
    "title": "픽업 서버 API",
    "description": "픽업 서버 API 문서",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateMenuItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "아메리카노",
            "description": "메뉴 이름"
          },
          "description": {
            "type": "string",
            "example": "진한 에스프레소에 물을 더한 음료",
            "description": "메뉴 설명"
          },
          "price": {
            "type": "number",
            "example": 4000,
            "description": "메뉴 가격"
          },
          "discountedPrice": {
            "type": "number",
            "example": 3500,
            "description": "할인된 가격"
          },
          "categoryId": {
            "type": "number",
            "example": 1,
            "description": "메뉴 카테고리 ID"
          },
          "image": {
            "type": "string",
            "description": "메뉴 이미지 파일",
            "format": "binary"
          },
          "preparationTime": {
            "type": "number",
            "example": 5,
            "description": "준비 시간 (분)"
          },
          "isPopular": {
            "type": "boolean",
            "example": true,
            "description": "인기 메뉴 여부"
          },
          "isNew": {
            "type": "boolean",
            "example": true,
            "description": "신메뉴 여부"
          },
          "isRecommended": {
            "type": "boolean",
            "example": true,
            "description": "추천 메뉴 여부"
          }
        },
        "required": [
          "name",
          "price"
        ]
      },
      "UpdateMenuItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "아이스 아메리카노",
            "description": "메뉴 이름"
          },
          "description": {
            "type": "string",
            "example": "차가운 에스프레소에 물을 더한 음료",
            "description": "메뉴 설명"
          },
          "price": {
            "type": "number",
            "example": 4500,
            "description": "메뉴 가격"
          },
          "discountedPrice": {
            "type": "number",
            "example": 4000,
            "description": "할인된 가격"
          },
          "categoryId": {
            "type": "number",
            "example": 1,
            "description": "메뉴 카테고리 ID"
          },
          "imageUrl": {
            "type": "string",
            "example": "iced_coffee.jpg",
            "description": "메뉴 이미지"
          },
          "preparationTime": {
            "type": "number",
            "example": 3,
            "description": "준비 시간 (분)"
          },
          "isAvailable": {
            "type": "boolean",
            "example": true,
            "description": "판매 가능 여부"
          },
          "isPopular": {
            "type": "boolean",
            "example": true,
            "description": "인기 메뉴 여부"
          },
          "isNew": {
            "type": "boolean",
            "example": false,
            "description": "신메뉴 여부"
          },
          "isRecommended": {
            "type": "boolean",
            "example": true,
            "description": "추천 메뉴 여부"
          },
          "stockQuantity": {
            "type": "number",
            "example": 30,
            "description": "재고 수량"
          },
          "image": {
            "type": "string",
            "description": "메뉴 이미지 파일",
            "format": "binary"
          }
        }
      },
...
      "UpdateOptionItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "옵션 항목 이름"
          },
          "price": {
            "type": "number",
            "description": "가격"
          },
          "isAvailable": {
            "type": "boolean",
            "example": true,
            "description": "옵션 가용 상태"
          },
          "displayOrder": {
            "type": "number",
            "description": "표시 순서"
          }
        }
      },
...
    }
  }
}